Had to add a halfbit init and fixed letter A
The first message is assumed to be 8 bit. so added halfBitInit().
Also, the letter A was encoded backwards in test.

In order to finish, figure out how to encode all letters in ASCII. There
should be a library on how to do that.

For enhanced visual pleasure, add in a string print function

For better performance, Reduce all of the delays inside of the write

For shorter delay after clearing and writing, make the write command
query the busy flag